+++ package.json
@@ -35,6 +35,7 @@
     "@types/express": "^4.17.14",
     "@types/html-minifier-terser": "^7.0.0",
     "@types/is-ci": "^3.0.0",
+    "@types/lodash": "^4.14.185",
     "@types/lodash.once": "^4.1.7",
     "@types/minimist": "^1.2.2",
     "@types/mocha": "^9.1.1",
