+++ packages/engineer/package.json
@@ -24,6 +24,7 @@
     "@wixc3/engine-runtime-node": "^29.0.1",
     "@wixc3/engine-scripts": "^29.0.1",
     "commander": "^9.4.0",
+    "lodash": "^4.17.21",
     "open": "^8.4.0",
     "socket.io": "^4.5.2",
     "webpack-dev-middleware": "^5.3.3"
