+++ packages/engineer/src/feature/dev-server.feature.ts
@@ -1,50 +1,9 @@
-import type io from 'socket.io';
 import type webpack from 'webpack';
-import { Feature, Service, Environment, COM, Config, TopLevelConfig, Slot } from '@wixc3/engine-core';
-import type { IExternalDefinition, TopLevelConfigProvider, LaunchEnvironmentMode } from '@wixc3/engine-runtime-node';
+import { Feature, Service, Environment, COM, Config, Slot } from '@wixc3/engine-core';
 import type { TargetApplication } from '../application-proxy-service';
-import type { Express } from 'express';
+import type { DevServerActions, DevServerConfig, ServerListeningHandler } from './dev-server.types';
 
 export const devServerEnv = new Environment('dev-server', 'node', 'single');
-
-export interface DevServerConfig {
-    httpServerPort: number;
-    featureName?: string;
-    singleFeature?: boolean;
-    configName?: string;
-    publicPath?: string;
-    title?: string;
-    favicon?: string;
-    publicConfigsRoute: string;
-    inspect: boolean;
-    autoLaunch: boolean;
-    nodeEnvironmentsMode?: LaunchEnvironmentMode;
-    basePath?: string;
-    mode: 'production' | 'development';
-    overrideConfig: TopLevelConfig | TopLevelConfigProvider;
-    defaultRuntimeOptions: Record<string, string | boolean>;
-    outputPath?: string;
-    externalFeatureDefinitions: IExternalDefinition[];
-    externalFeaturesPath?: string;
-    serveExternalFeaturesPath?: boolean;
-    featureDiscoveryRoot?: string;
-    socketServerOptions?: Partial<io.ServerOptions>;
-    webpackConfigPath?: string;
-    externalFeaturesRoute: string;
-    log?: boolean;
-}
-
-export interface DevServerActions {
-    close: () => Promise<void>;
-}
-
-export interface ServerListeningParams {
-    port: number;
-    host: string;
-    router: Express;
-}
-
-export type ServerListeningHandler = (params: ServerListeningParams) => void | Promise<void>;
 
 export default new Feature({
     id: 'buildFeature',
