+++ packages/engineer/src/feature/dev-server.types.ts
@@ -1,0 +1,43 @@
+import type { Express } from 'express';
+import type { IExternalDefinition, TopLevelConfigProvider, LaunchEnvironmentMode } from '@wixc3/engine-runtime-node';
+import type { TopLevelConfig } from '@wixc3/engine-core';
+import type io from 'socket.io';
+
+export interface DevServerConfig {
+    httpServerPort: number;
+    featureName?: string;
+    singleFeature?: boolean;
+    configName?: string;
+    publicPath?: string;
+    title?: string;
+    favicon?: string;
+    publicConfigsRoute: string;
+    inspect: boolean;
+    autoLaunch: boolean;
+    nodeEnvironmentsMode?: LaunchEnvironmentMode;
+    basePath?: string;
+    mode: 'production' | 'development';
+    overrideConfig: TopLevelConfig | TopLevelConfigProvider;
+    defaultRuntimeOptions: Record<string, string | boolean>;
+    outputPath?: string;
+    externalFeatureDefinitions: IExternalDefinition[];
+    externalFeaturesPath?: string;
+    serveExternalFeaturesPath?: boolean;
+    featureDiscoveryRoot?: string;
+    socketServerOptions?: Partial<io.ServerOptions>;
+    webpackConfigPath?: string;
+    externalFeaturesRoute: string;
+    log?: boolean;
+}
+
+export interface DevServerActions {
+    close: () => Promise<void>;
+}
+
+export interface ServerListeningParams {
+    port: number;
+    host: string;
+    router: Express;
+}
+
+export type ServerListeningHandler = (params: ServerListeningParams) => void | Promise<void>;
